jsonrpc: convert peer-to-peer-with-events test to forge
