+++ json-rpc/tests/integration_test.rs
@@ -63,116 +63,6 @@ pub struct Test {
 fn create_test_cases() -> Vec<Test> {
     vec![
         Test {
-            name: "peer to peer account transaction with events",
-            run: |env: &mut testing::Env| {
-                let prev_ledger_version = env.send("get_metadata", json!([])).diem_ledger_version;
-
-                let txn = env.transfer_coins((0, 0), (1, 0), 200000);
-                env.wait_for_txn(&txn);
-                let txn_hex = hex::encode(bcs::to_bytes(&txn).expect("bcs txn failed"));
-
-                let sender = &env.vasps[0].children[0];
-                let receiver = &env.vasps[1].children[0];
-
-                let resp = env.send(
-                    "get_account_transaction",
-                    json!([sender.address.to_string(), 0, true]),
-                );
-                let result = resp.result.unwrap();
-                let version = result["version"].as_u64().unwrap();
-                assert_eq!(
-                    true,
-                    version > prev_ledger_version && version <= resp.diem_ledger_version
-                );
-
-                let gas_used = result["gas_used"].as_u64().expect("exist as u64");
-                let txn_hash = Transaction::UserTransaction(txn.clone()).hash().to_hex();
-
-                let script = match txn.payload() {
-                    TransactionPayload::Script(s) => s,
-                    _ => unreachable!(),
-                };
-                let script_hash = diem_crypto::HashValue::sha3_256_of(script.code()).to_hex();
-                let script_bytes = hex::encode(bcs::to_bytes(script).unwrap());
-
-                assert_eq!(
-                    result,
-                    json!({
-                        "bytes": format!("00{}", txn_hex),
-                        "events": [
-                            {
-                                "data": {
-                                    "amount": {"amount": 200000_u64, "currency": "XUS"},
-                                    "metadata": "",
-                                    "receiver": format!("{:x}", &receiver.address),
-                                    "sender": format!("{:x}", &sender.address),
-                                    "type": "sentpayment"
-                                },
-                                "key": format!("0100000000000000{:x}", &sender.address),
-                                "sequence_number": 0,
-                                "transaction_version": version,
-                            },
-                            {
-                                "data": {
-                                    "amount": {"amount": 200000_u64, "currency": "XUS"},
-                                    "metadata": "",
-                                    "receiver": format!("{:x}", &receiver.address),
-                                    "sender": format!("{:x}", &sender.address),
-                                    "type": "receivedpayment"
-                                },
-                                "key": format!("0000000000000000{:x}", &receiver.address),
-                                "sequence_number": 1,
-                                "transaction_version": version
-                            }
-                        ],
-                        "gas_used": gas_used,
-                        "hash": txn_hash,
-                        "transaction": {
-                            "chain_id": 4,
-                            "expiration_timestamp_secs": txn.expiration_timestamp_secs(),
-                            "gas_currency": "XUS",
-                            "gas_unit_price": 0,
-                            "max_gas_amount": 1000000,
-                            "public_key": sender.public_key.to_string(),
-                            "secondary_signers": [],
-                            "secondary_signature_schemes": [],
-                            "secondary_signatures": [],
-                            "secondary_public_keys": [],
-                            "script": {
-                                "type": "peer_to_peer_with_metadata",
-                                "type_arguments": [
-                                    "XUS"
-                                ],
-                                "arguments": [
-                                    format!("{{ADDRESS: {:?}}}", &receiver.address),
-                                    "{U64: 200000}",
-                                    "{U8Vector: 0x}",
-                                    "{U8Vector: 0x}"
-                                ],
-                                "code": hex::encode(script.code()),
-                                "amount": 200000,
-                                "currency": "XUS",
-                                "metadata": "",
-                                "metadata_signature": "",
-                                "receiver": format!("{:x}", &receiver.address),
-                            },
-                            "script_bytes": script_bytes,
-                            "script_hash": script_hash,
-                            "sender": format!("{:x}", &sender.address),
-                            "sequence_number": 0,
-                            "signature": hex::encode(txn.authenticator().sender().signature_bytes()),
-                            "signature_scheme": "Scheme::Ed25519",
-                            "type": "user"
-                        },
-                        "version": version,
-                        "vm_status": {"type": "executed"}
-                    }),
-                    "{:#}",
-                    result,
-                );
-            },
-        },
-        Test {
             name: "peer to peer account failed error explanations",
             run: |env: &mut testing::Env| {
                 env.allow_execution_failures(|env: &mut testing::Env| {
